Move findDestructuredProps to language-service
